@@ -5,25 +5,17 @@ import java.io.InputStreamReader;
 import java.net.URL;
 import java.util.Objects;
 
+import me.MrGraycat.eGlow.Addon.BlockWars.BlockWarsAddon;
 import me.MrGraycat.eGlow.Addon.Internal.AdvancedGlowVisibilityAddon;
-import me.MrGraycat.eGlow.Addon.TabList.TabListAddon;
 import org.bstats.bukkit.Metrics;
 import org.bstats.charts.SimplePie;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
-import org.bukkit.plugin.Plugin;
 import org.bukkit.plugin.java.JavaPlugin;
 import org.bukkit.scheduler.BukkitRunnable;
 
 import me.MrGraycat.eGlow.API.EGlowAPI;
-import me.MrGraycat.eGlow.Addon.Citizens.CitizensAddon;
-import me.MrGraycat.eGlow.Addon.Disguises.IDisguiseAddon;
-import me.MrGraycat.eGlow.Addon.Disguises.LibDisguiseAddon;
-import me.MrGraycat.eGlow.Addon.LuckPermsAddon;
 import me.MrGraycat.eGlow.Addon.PlaceholderAPIAddon;
-import me.MrGraycat.eGlow.Addon.TAB.Listeners.EGlowTABListenerUniv;
-import me.MrGraycat.eGlow.Addon.TAB.TABAddon;
-import me.MrGraycat.eGlow.Addon.VaultAddon;
 import me.MrGraycat.eGlow.Command.EGlowCommand;
 import me.MrGraycat.eGlow.Config.EGlowCustomEffectsConfig;
 import me.MrGraycat.eGlow.Config.EGlowMainConfig;
@@ -44,13 +36,6 @@ public class EGlow extends JavaPlugin {
 
 	//Addons
 	private AdvancedGlowVisibilityAddon glowAddon;
-	private CitizensAddon citizensAddon;
-	private IDisguiseAddon iDisguiseAddon;
-	private LibDisguiseAddon libDisguiseAddon;
-	private TABAddon tabAddon;
-	private TabListAddon tablistAddon;
-	private LuckPermsAddon lpAddon;
-	private VaultAddon vaultAddon;
 	private Metrics metrics;
 	
 	@Override
@@ -71,6 +56,8 @@ public class EGlow extends JavaPlugin {
 			checkForUpdates();
 			runAddonHooks();
 			runPlayerCheckOnEnable();
+
+			BlockWarsAddon.load();
 		} else {
 			ChatUtil.sendToConsole("Disabling eGlow! Your server version is not compatible! (" + DebugUtil.getServerVersion() + ")", false);
 			getServer().getPluginManager().disablePlugin(this);
@@ -81,10 +68,6 @@ public class EGlow extends JavaPlugin {
 	public void onDisable() {
 		if (getAdvancedGlowVisibility() != null) {
 			getAdvancedGlowVisibility().shutdown();
-		}
-
-		if (getLPAddon() != null) {
-			getLPAddon().unload();
 		}
 
 		runPlayerCheckOnDisable();
@@ -110,41 +93,10 @@ public class EGlow extends JavaPlugin {
 		new BukkitRunnable() {
 			@Override
 			public void run() {
-				setMetricsAddon(new Metrics(getInstance(), 9468));
-
-				getMetricsAddon().addCustomChart(new SimplePie("custom_glow_visibility_usage", () -> (MainConfig.ADVANCED_GLOW_VISIBILITY_ENABLE.getBoolean()) ? "Yes" : "No"));
-				getMetricsAddon().addCustomChart(new SimplePie("database_type", () -> (MainConfig.MYSQL_ENABLE.getBoolean()) ? "MySQL" : "SQLite"));
-
 				if (MainConfig.ADVANCED_GLOW_VISIBILITY_ENABLE.getBoolean() && getAdvancedGlowVisibility() == null)
 					setAdvancedGlowVisibility(new AdvancedGlowVisibilityAddon());
 				if (DebugUtil.pluginCheck("PlaceholderAPI"))
 					new PlaceholderAPIAddon();
-				if (DebugUtil.pluginCheck("Vault"))
-					setVaultAddon(new VaultAddon());
-				if (DebugUtil.pluginCheck("Citizens") && getCitizensAddon() == null)
-					setCitizensAddon(new CitizensAddon());
-				if (DebugUtil.pluginCheck("iDisguise"))
-					setIDisguiseAddon(new IDisguiseAddon());
-				if (DebugUtil.pluginCheck("LibsDisguises"))
-					setLibDisguiseAddon(new LibDisguiseAddon());
-				if (DebugUtil.pluginCheck("TAB")) {
-					try {
-						Plugin TAB_Plugin = DebugUtil.getPlugin("TAB");
-						
-						if (TAB_Plugin != null && TAB_Plugin.getClass().getName().startsWith("me.neznamy.tab"))
-							setTABAddon(new TABAddon(TAB_Plugin));
-					} catch (NoClassDefFoundError e) {			
-						ChatUtil.sendToConsole("&cWarning&f! &cThis version of eGlow requires TAB 3.1.0 or higher!", true);
-					}
-				} else if (DebugUtil.pluginCheck("TabList")) {
-					setTablistAddon(new TabListAddon());
-				}
-				
-				EGlow.getInstance().getServer().getPluginManager().registerEvents(new EGlowTABListenerUniv(), getInstance());
-				
-				if (DebugUtil.pluginCheck("LuckPerms")) {
-					setLPAddon(new LuckPermsAddon());
-				}
 			}
 		}.runTask(this);
 	}
@@ -206,34 +158,6 @@ public class EGlow extends JavaPlugin {
 	public void setAdvancedGlowVisibility(AdvancedGlowVisibilityAddon glowAddon) {
 		this.glowAddon = glowAddon;
 	}
-
-	private void setCitizensAddon(CitizensAddon citizensAddon) {
-		this.citizensAddon = citizensAddon;
-	}
-	
-	private void setIDisguiseAddon(IDisguiseAddon iDisguiseAddon) {
-		this.iDisguiseAddon = iDisguiseAddon;
-	}
-	
-	private void setLibDisguiseAddon(LibDisguiseAddon libDisguiseAddon) {
-		this.libDisguiseAddon = libDisguiseAddon;
-	}
-	
-	private void setTABAddon(TABAddon tabAddon) {
-		this.tabAddon = tabAddon;
-	}
-
-	private void setTablistAddon(TabListAddon tablistAddon) {
-		this.tablistAddon = tablistAddon;
-	}
-	
-	private void setLPAddon(LuckPermsAddon lpAddon) {
-		this.lpAddon = lpAddon;
-	}
-	
-	private void setVaultAddon(VaultAddon vaultAddon) {
-		this.vaultAddon = vaultAddon;
-	}
 	
 	//Getter
 	public static EGlow getInstance() {
@@ -254,33 +178,5 @@ public class EGlow extends JavaPlugin {
 
 	public AdvancedGlowVisibilityAddon getAdvancedGlowVisibility() {
 		return this.glowAddon;
-	}
-
-	public CitizensAddon getCitizensAddon() {
-		return this.citizensAddon;
-	}
-	
-	public IDisguiseAddon getIDisguiseAddon() {
-		return this.iDisguiseAddon;
-	}
-	
-	public LibDisguiseAddon getLibDisguiseAddon() {
-		return this.libDisguiseAddon;
-	}
-	
-	public TABAddon getTABAddon() {
-		return this.tabAddon;
-	}
-
-	public TabListAddon getTablistAddon() {
-		return this.tablistAddon;
-	}
-	
-	public LuckPermsAddon getLPAddon() {
-		return this.lpAddon;
-	}
-	
-	public VaultAddon getVaultAddon() {
-		return this.vaultAddon;
 	}
 }
